do not force the Response to have a content-type header
Not all http responses require a content-type. For example, 304 not
modified. It seems unlikely that setting the content-type to text/html
when no content-type header is found is seldom going to actually fix
something for the better. Most methods of creating html will already
correctly set the content-type header.

This should close this bug:

http://code.google.com/p/happstack/issues/detail?id=122