removal HaXml dependency/support This package removes support for and use of HaXml. This should make it easier to install Happstack, and to get it Debian (which ships HaXml 1.13) and Arch (which uses > 1.21).
We lose a few features a result, but probably not anything that people are using:

 1. ToMessage instance for HaXml Element type
 2. some functionality in the happpstack-data XML code

The HaXml code still existings in the repo, but it is not built by default, so if you really need these features, you can re-enable them locally. Also, be sure to let us know, so that we can work to restore them.