Don't unescape +'s in path segments in happstack.
This is not spec, and leads to problems with Apache rewrite rules.
Apache unescapes %2B to +, and doesn't re-escape, because it isn't a
special character. Happstack then makes it a space.

Doesn't unescape '+' characters in path segments
anymore. It still unescapes them in query strings. This is according
to the spec. I ran into this because apache decodes encoded '+'
characters for rewrite rules, and doesn't re-encode them anymore
(since they don't need to be). So they were passed to happstack, which
decoded them again. I renamed the current function unescape to
unescapeQS, since I felt query strings were the special case. On
second thought, keeping the old function and using a separate one for
paths might be cleaner, since it doesn't break the old API.